moved functions to conftest
